+++ templates/junction/new/_inject-parent-service-counterparty-clp-left.ejs.t
@@ -1,7 +1,0 @@
----
-to: "<%= exposeOnParent.left ? parentServicePathLeft : '' %>"
-inject: true
-after: "from '@nestjs/common';"
-skip_if: "from '<%= rightEntityImportFromLeft %>'"
----
-import type { <%= rightEntityPascal %> } from '<%= rightEntityImportFromLeft %>';