+++ templates/junction/new/_inject-parent-service-counterparty-clp-right.ejs.t
@@ -1,7 +1,0 @@
----
-to: "<%= exposeOnParent.right ? parentServicePathRight : '' %>"
-inject: true
-after: "from '@nestjs/common';"
-skip_if: "from '<%= leftEntityImportFromRight %>'"
----
-import type { <%= leftEntityPascal %> } from '<%= leftEntityImportFromRight %>';